requirements.txt
streamlit>=1.30.0
azure-identity>=1.15.0
openai>=1.70.0
requests>=2.31.0
plotly>=5.18.0
pandas>=2.0.0
